Time: 1692 ms (6.56%), Space: 10.8 MB (22.34%) - LeetHub

+++ 923-3sum-with-multiplicity/923-3sum-with-multiplicity.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int threeSumMulti(vector<int>& arr, int target) {
         int n = arr.size();
-        int mod = (int) 1e9+7;
+        int mod = 1e9+7;
         int ans = 0, i=0, t;
         unordered_map<int, int> map;
         for(int i=0; i<n; i++){
